refactor source/s3v2
